tweaked code for pi 4

+++ src/LocalizationServer.py
@@ -5,6 +5,7 @@ import asyncio
 import glob
 import threading
 import serial
+from os.path import exists
 # from imu import getIMUData
 
 STORAGE_LOCATIONS = glob.glob("/media/pi/*")
@@ -32,15 +33,9 @@ def __async__data_write(): # usb/sd entrypoint
     asyncio.run(handle_data())
 
 async def handle_data():
-    # TODO: Grab data from IMU/GPU (likely from a wrapper function). 
-    #       Each communication process should attempt to recieve the data from the IMU/GPS
-    # TODO: send data through USB!
-    # event_loop = asyncio.get_event_loop()
-    # event_loop.create_task(store_to_sd())
-    # event_loop.run_forever()
     while True:
         localization_data = "getData() "
-        await store_to_sd(localization_data)
+        await write_to_sd(localization_data)
         await sendSerial(localization_data)
 
 async def getData():
@@ -49,23 +44,25 @@ async def getData():
     return imu_data
 
 async def sendSerial(message):
-    #port = serial.Serial('/dev/serial1', 9600) # TODO: make sure this doesn't need to be changed!!
+    #port = serial.Serial('/dev/ttyAMA1', 9600) # TODO: make sure this doesn't need to be changed!!
     #byte_message = bytes(message, 'utf-8')
     #port.write(byte_message)
     #port.close()
     print("Sent Data via Serial")
-    
-async def store_to_sd(data):
-    await write_to_sd(data)
 
 async def write_to_sd(data):
     if(len(STORAGE_LOCATIONS) == 0):
         return False
     else:
-        PATH = STORAGE_LOCATIONS[0]+"/LocalizationData.csv" #TODO: Change to CSV file extension
-        f = open(PATH, "w")
-        f.write(str(data + "\n")) #TODO: Might get rid of the /n for testing
-        f.close()
+        PATH = STORAGE_LOCATIONS[0]+"/LocalizationData.csv"
+        if(exists(PATH)):
+            f = open(PATH, "a")
+            f.write(str(data + "\n"))
+            f.close()
+        else:
+            f = open(PATH, "w")
+            f.write(str(data + "\n"))
+            f.close()
         return True
